+++ include/fred.h
@@ -15,7 +15,7 @@ class Fred : public Character {
 
         Fred(double x, double y);
 
-        void damage(int d) override;
+        void damage(float d) override;
 };
 
 #endif
